Sound and articulated sound creation for delegating playing to the class not the block. The MIDIChannel class support various simple actions to seperate the midi data from the actions for a better seperation of concern.

+++ src/main/java/cas/cs4tb3/mellowd/TimingEnvironment.java
@@ -46,9 +46,9 @@ public class TimingEnvironment {
 
     //The numerator and denominator for the time signature gives the synthesiser some hints
     //on what are the down and up beats. The bpm is the tempo.
-    public TimingEnvironment(byte numerator, byte denominator, int bpm) {
-        this.timeSigNum = numerator;
-        this.timeSigDen = denominator;
+    public TimingEnvironment(int numerator, int denominator, int bpm) {
+        this.timeSigNum = (byte) numerator;
+        this.timeSigDen = (byte) denominator;
         this.ppqn = DEFAULT_PPQN;
         this.bpm = bpm;
     }
